requirements.txt
dataclasses
dnspython
tabulate
pycryptodome>=3.4
requests
validators
python-dateutil
ecdsa
